- moves personalization to feed package, separate interest into its own package
