+++ app/src/main/AndroidManifest.xml
@@ -422,7 +422,7 @@
 
         <activity android:name=".feed.onboarding.ExploreFeedBuildingActivity"/>
 
-        <activity android:name=".onboarding.personalization.PersonalizationActivity"/>
+        <activity android:name=".feed.personalization.PersonalizationActivity"/>
 
         <provider
             android:name=".WikipediaFileProvider"
