+++ app/src/main/java/org/wikipedia/feed/onboarding/ExploreFeedUpdatePromptActivity.kt
@@ -27,7 +27,7 @@ import org.wikipedia.compose.components.OnboardingListItem
 import org.wikipedia.compose.components.TwoButtonBottomBar
 import org.wikipedia.compose.theme.BaseTheme
 import org.wikipedia.compose.theme.WikipediaTheme
-import org.wikipedia.onboarding.personalization.PersonalizationActivity
+import org.wikipedia.feed.personalization.PersonalizationActivity
 import org.wikipedia.settings.Prefs
 import org.wikipedia.theme.Theme
