+++ app/src/main/java/org/wikipedia/feed/personalization/interest/InterestSelectionState.kt
@@ -1,4 +1,4 @@
-package org.wikipedia.onboarding.personalization
+package org.wikipedia.feed.personalization.interest
 
 import org.wikipedia.page.PageTitle
